fastapi/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
langchain
openai
lark
tiktoken
PyPDF2
firebase-admin
python-decouple